ai_hr_system/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
pdfplumber>=0.10.0
python-docx>=0.8.11
spacy>=3.5.0
python-multipart>=0.0.6
sentence-transformers>=2.2.0,<3.0.0
huggingface-hub>=0.20.0,<1.0.0
numpy>=1.24.0
aiogram==3.1.1
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
